+++ mchap/application/arguments.py
@@ -348,7 +348,7 @@ report = Parameter(
         default=[],
         help=(
             "Extra fields to report within the output VCF: "
-            "AFPR = prior allele frequencies; "
+            "AFPRIOR = prior allele frequencies; "
             "AFP = posterior mean allele frequencies; "
             "GP = genotype posterior probabilities; "
             "GL = genotype likelihoods."