+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 New Features:
 - Mask reference allele when it is only reported to satisfy VCF spec #146
-- Optionally report prior allele frequencies in `AFPR` field
+- Optionally report prior allele frequencies in `AFPRIOR` field
 - Filtering for some edge cases where genotypes should not be reported (`AF0` and `NOA`)
 
 Bug Fixes:
@@ -14,7 +14,7 @@ Bug Fixes:
 
 VCF Changes:
 - Added `REFMASKED` info flag to indicate reference allele is amsked and should be ignored
-- Added `AFPR` infor filed to indicate prior allele frequencies
+- Added `AFPRIOR` infor filed to indicate prior allele frequencies
 - Added `NOA` filter to indicate loci where no alleles were observed (e.g., masked reference only)
 - Added `AF0` filter to indicate invalid prior allele frequencies in which all frequencies were zero